perf: 主要4ページの読み込み速度を大幅改善
- 顧客・予約・売上ページをServer Component化（クライアント側のauth+fetch往復を排除）
- ダッシュボードのクエリを11→8に削減（月間売上3クエリ→1 RPC、在庫アラートを並列化）
- ミドルウェアのgetUser()→getSession()に変更（API問い合わせ→Cookie読み取りで100-200ms短縮）
- DateNavigator共通コンポーネント抽出
